requirements.txt
django-extensions==3.2.1
djangorestframework==3.14.0